pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scorpio-pipe"
requires-python = ">=3.10"
readme = "README.md"
# NOTE: setuptools (as used in CI) validates PEP 621 `license` as a string.
# We ship the license file via `project.license-files` (setuptools>=77).
license = "BSD-3-Clause"
license-files = ["LICENSE"]
dependencies = [
  "astropy>=6",
  "numpy",
  "pandas",
  "pyyaml",
  "pydantic>=2",
  "doit",
  "rich",
  "loguru",
  "matplotlib",
  # Scientific stack: do not skimp on accuracy.
  "scipy>=1.10",
  # Mandatory cosmic-ray cleaner engine (P1): L.A.Cosmic family.
  "astroscrappy>=1.2",
]

dynamic = ["version"]
[project.optional-dependencies]
# GUI stack for interactive LineID (and embedded PDF viewing)
gui = [
  "PySide6>=6.5",
  "PySide6-Addons>=6.5",
  "pyqtgraph>=0.13",
  "pymupdf>=1.23"
]

# Recommended scientific stack (better peak finding, etc.)
science = [
  # SciPy is now part of the core dependencies (v5.40.2+).
  # This extra remains for backward compatibility with older install guides.
  "scipy>=1.10"
]

# Dev tooling (tests / lint). Optional.
dev = [
  "pytest>=8",
  "ruff>=0.7"
]

[project.scripts]
scorpio-pipe = "scorpio_pipe.cli:main"

[project.gui-scripts]
# Windows: creates a no-console launcher (pythonw) so the UI can be started by double click.
scorpio-ui = "scorpio_pipe.ui.launcher_app:main"
scorpipe = "scorpio_pipe.ui.launcher_app:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.dynamic]
version = { attr = "scorpio_pipe.version.__version__" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
scorpio_pipe = ["resources/**/*"]
